Module external_storage_export::export

To use External storage with protobufs as an application, import this module. external_storage contains the actual library code Cloud provider backends are under components/cloud

Structs

BlobStore

EncryptedExternalStorage

GCSConfig

GCSStorage

LocalStorage
A storage saves files in local file system.

NoopStorage
A storage saves files into void. It is mainly for test use.

S3Config

S3Storage

Enums

Backend

Traits

ExternalStorage
An abstraction of an external storage.

Functions

bad_backend

bad_storage_backend

blob_store

create_backend

create_backend_inner
Create a new storage from the given storage backend description.

create_storage

create_storage_no_client

make_cloud_backend

make_gcs_backend

make_local_backend

make_noop_backend
Creates a noop StorageBackend.

make_s3_backend

read_external_storage_into_file